signal23-ai/requirements.txt
# API Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1

# LLM & RAG
langchain==0.3.8
langchain-community==0.3.8
langchain-core==0.3.21
langchain-ollama==0.2.0
langchain-openai==0.2.9
langchain-text-splitters==0.3.2

# Vector Storage
pgvector==0.2.4
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9

# Notion Integration
notion-client==2.2.1

# Utils
python-dotenv==1.0.0
pydantic>=2.7.4,<3.0.0  # Updated to match langchain's requirements
boto3==1.35.69
botocore==1.35.69
httpx-sse==0.4.0